# src/api/radar/openApi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Signal K Radar API",
    "version": "2.0.0",
    "description": "API for managing marine radar devices"
  },
  "paths": {
...
    "/signalk/v2/api/vessels/self/radars/{id}": {
      "get": {
        "tags": ["radar"],
        "summary": "Get radar info",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Radar information",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RadarInfo" }
              }
            }
          },
          "404": { "description": "Radar not found" }
        }
      },
      "put": {
        "tags": ["radar"],
        "summary": "Update radar controls",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RadarControls" }
            }
          }
        },
        "responses": {
          "200": { "description": "Controls updated" },
          "403": { "description": "Unauthorized" },
          "404": { "description": "Radar not found" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "RadarInfo": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "description": "Unique radar identifier" },
          "name": { "type": "string", "description": "Display name" },
          "brand": { "type": "string", "description": "Manufacturer/brand" },
          "status": {
            "type": "string",
            "enum": ["off", "standby", "transmit", "warming"],
            "description": "Current operational status"
          },
          "spokesPerRevolution": {
            "type": "integer",
            "description": "Number of spokes per full rotation"
          },
          "maxSpokeLen": {
            "type": "integer",
            "description": "Maximum spoke length in samples"
          },
          "range": {
            "type": "number",
            "description": "Current range in meters"
          },
          "controls": { "$ref": "#/components/schemas/RadarControls" },
          "legend": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/LegendEntry" }
          },
          "streamUrl": {
            "type": "string",
            "description": "WebSocket URL for spoke stream. If absent, use /radars/{id}/stream"
          }
        },
        "required": [
          "id",
          "name",
          "status",
          "spokesPerRevolution",
          "maxSpokeLen",
          "range",
          "controls"
        ]
      },
      "RadarControls": {
        "type": "object",
        "properties": {
          "gain": { "$ref": "#/components/schemas/ControlValue" },
          "sea": { "$ref": "#/components/schemas/ControlValue" },
          "rain": {
            "type": "object",
            "properties": { "value": { "type": "number" } }
          }
        },
        "required": ["gain"]
      },
      "ControlValue": {
        "type": "object",
        "properties": {
          "auto": { "type": "boolean" },
          "value": { "type": "number" }
        },
        "required": ["auto", "value"]
      },
      "LegendEntry": {
        "type": "object",
        "properties": {
          "color": { "type": "string" },
          "label": { "type": "string" },
          "minValue": { "type": "number" },
          "maxValue": { "type": "number" }
        },
        "required": ["color", "label"]
      }
    }
  }
}